rename runtime fields
